+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 @samber: I sometimes forget to update this file. Ping me on [Twitter](https://twitter.com/samuelberthe) or open an issue in case of error. We need to keep a clear changelog for easier lib upgrade.
 
+## 1.34.1 (2022-11-xx)
+
+Adding:
+- lo.BufferWithTimeout (alias to lo.BatchWithTimeout)
+- lo.Buffer (alias to lo.Batch)
+
+Deprecation:
+- lo.BatchWithTimeout
+- lo.Batch
+
 ## 1.34.0 (2022-11-12)
 
 Improving: